Do the arithmetic in long where the result is a long
Sonar java:S2184.

+++ DEV/org.openl.commons/src/org/openl/util/fast/FastStringReader.java
@@ -102,7 +102,7 @@ public class FastStringReader extends Reader {
             return 0;
         }
         // Bound skip by beginning and end of the source
-        var n = Math.min(length - next, ns);
+        var n = Math.min((long) length - next, ns);
         n = Math.max(-next, n);
         next += n;
         return n;
